requirements.txt
altgraph==0.17.2
autopep8==1.5.2
certifi==2021.5.30
charset-normalizer==2.0.6
future==0.18.2
idna==3.2
importlib-metadata==4.8.1
numpy==1.21.2
pandas==1.3.3
pefile==2021.9.3
Pillow==8.3.2
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pytesseract==0.3.8
python-dateutil==2.8.2
pytz==2021.1
pywin32-ctypes==0.2.0
requests==2.26.0
selenium==3.141.0
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.6
wincertstore==0.2
zipp==3.6.0
